% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLoadFile.fn.R
\name{GetLoadFile}
\alias{GetLoadFile}
\title{Get Load File}
\usage{
GetLoadFile(dir.path, type = "summary")
}
\arguments{
\item{type}{type of output to retrieve (i.e., summary, cohort, or nest)}

\item{file}{path to run directory containing SJRSim simulation run output}
}
\value{

}
\description{
Get Load File
}
\examples{

}
