% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuvenilesOutByCohort.fn.R
\name{JuvenilesOutByCohort}
\alias{JuvenilesOutByCohort}
\title{Juvenile Cohort Summaires at Exit}
\usage{
JuvenilesOutByCohort(run.dir, river = NULL)
}
\arguments{
\item{run.dir}{path to SJRSim run output directory.}

\item{river}{optional names of birth rivers to return, if not specified all rivers returned.}
}
\description{
Juvenile Cohort Summaires at Exit
}
