% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuvenilesOutByStage.fn.R
\name{JuvenilesOutByLifeStage}
\alias{JuvenilesOutByLifeStage}
\title{Juveniles Exiting by Life Stage}
\usage{
JuvenilesOutByLifeStage(run.dir, river)
}
\arguments{
\item{run.dir}{path to SJRSim run output directory.}

\item{river}{}
}
\value{
data frame with
}
\description{
Juveniles Exiting by Life Stage
}
