% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadHEC.R
\name{ReadHEC}
\alias{ReadHEC}
\title{Read HEC flow/temp file}
\usage{
ReadHEC(file, wide.format = FALSE, date.format = "\%m/\%d/\%Y",
  debug.mode = FALSE)
}
\arguments{
\item{file}{}

\item{wide.format}{}

\item{date.format}{}

\item{debug.mode}{}
}
\description{
Read HEC flow/temp file
}
