% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spawners.fn.R
\name{Spawners}
\alias{Spawners}
\title{Retreive SJRSim Spawners Output}
\usage{
Spawners(run.dir, river = NULL, incl.historic = T)
}
\arguments{
\item{incl.historic}{}

\item{file}{file path to the NAME_Summary.csv}
}
\description{
Retreive SJRSim Spawners Output
}
\details{
The \emph{[NAME]_Summary.csv} file contains a number of summary report

Summary Report of Fish Counts
Summary Report Of Egg Mortality
Summary Report Of Fish Mortality by River of Birth or Hatchery Release
Manager's Summary Report
}
