% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteHEC.R
\name{WriteSalSimHEC}
\alias{WriteSalSimHEC}
\title{Write SalSim/SJRSim HEC file}
\usage{
WriteSalSimHEC(flowtemp, header = NULL, river = NULL, filename)
}
\arguments{
\item{flowtemp}{- flowtemp data should be in wide format.  Column names are those returen from readHEC}

\item{river}{- character string indicating the river name, options include: TUOLUMNE', 'STANISLAUS', 'SAN JOAQUIN', 'MERCED', 'FRIANT'}

\item{filename}{- File name to be written to.}
}
\value{
No values returned, rather a system file is written.
}
\description{
Write SalSim/SJRSim HEC file
}
\examples{

}
